+++ crates/scheduler/src/call/crypto/ccm.rs
@@ -17,6 +17,7 @@ use wasefire_applet_api::crypto::ccm::{self as api, Api};
 use wasefire_board_api::crypto::aead::{Api as _, Array};
 use wasefire_board_api::{self as board, Api as Board, Support};
 
+use crate::applet::store::MemoryApi;
 use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {