+++ crates/stub/src/crypto/hash.rs
@@ -56,12 +56,12 @@ impl Contexts {
 }
 
 #[no_mangle]
-unsafe extern "C" fn chs(_: api::is_supported::Params) -> api::is_supported::Results {
+unsafe extern "C" fn env_chs(_: api::is_supported::Params) -> api::is_supported::Results {
     api::is_supported::Results { supported: 1 }
 }
 
 #[no_mangle]
-unsafe extern "C" fn chi(params: api::initialize::Params) -> api::initialize::Results {
+unsafe extern "C" fn env_chi(params: api::initialize::Params) -> api::initialize::Results {
     let api::initialize::Params { algorithm } = params;
     let context = match api::Algorithm::from(algorithm) {
         api::Algorithm::Sha256 => Context::Sha256(Sha256::default()),
@@ -72,7 +72,7 @@ unsafe extern "C" fn chi(params: api::initialize::Params) -> api::initialize::Re
 }
 
 #[no_mangle]
-unsafe extern "C" fn chu(params: api::update::Params) {
+unsafe extern "C" fn env_chu(params: api::update::Params) {
     let api::update::Params { id, data, length } = params;
     let data = unsafe { std::slice::from_raw_parts(data, length) };
     match CONTEXTS.lock().unwrap().get(id) {
@@ -83,7 +83,7 @@ unsafe extern "C" fn chu(params: api::update::Params) {
 }
 
 #[no_mangle]
-unsafe extern "C" fn chf(params: api::finalize::Params) -> api::finalize::Results {
+unsafe extern "C" fn env_chf(params: api::finalize::Params) -> api::finalize::Results {
     let api::finalize::Params { id, digest } = params;
     let digest = |length| unsafe { std::slice::from_raw_parts_mut(digest, length) };
     match CONTEXTS.lock().unwrap().take(id) {
@@ -95,12 +95,14 @@ unsafe extern "C" fn chf(params: api::finalize::Params) -> api::finalize::Result
 }
 
 #[no_mangle]
-unsafe extern "C" fn cht(_: api::is_hmac_supported::Params) -> api::is_hmac_supported::Results {
+unsafe extern "C" fn env_cht(_: api::is_hmac_supported::Params) -> api::is_hmac_supported::Results {
     api::is_hmac_supported::Results { supported: 1 }
 }
 
 #[no_mangle]
-unsafe extern "C" fn chj(params: api::hmac_initialize::Params) -> api::hmac_initialize::Results {
+unsafe extern "C" fn env_chj(
+    params: api::hmac_initialize::Params,
+) -> api::hmac_initialize::Results {
     let api::hmac_initialize::Params { algorithm, key, key_len } = params;
     let key = unsafe { std::slice::from_raw_parts(key, key_len) };
     let context: Result<Context, InvalidLength> = try {
@@ -117,7 +119,7 @@ unsafe extern "C" fn chj(params: api::hmac_initialize::Params) -> api::hmac_init
 }
 
 #[no_mangle]
-unsafe extern "C" fn chv(params: api::hmac_update::Params) {
+unsafe extern "C" fn env_chv(params: api::hmac_update::Params) {
     let api::hmac_update::Params { id, data, length } = params;
     let data = unsafe { std::slice::from_raw_parts(data, length) };
     match CONTEXTS.lock().unwrap().get(id) {
@@ -128,7 +130,7 @@ unsafe extern "C" fn chv(params: api::hmac_update::Params) {
 }
 
 #[no_mangle]
-unsafe extern "C" fn chg(params: api::hmac_finalize::Params) -> api::hmac_finalize::Results {
+unsafe extern "C" fn env_chg(params: api::hmac_finalize::Params) -> api::hmac_finalize::Results {
     let api::hmac_finalize::Params { id, hmac } = params;
     let hmac = |length| unsafe { std::slice::from_raw_parts_mut(hmac, length) };
     match CONTEXTS.lock().unwrap().take(id) {
@@ -140,12 +142,12 @@ unsafe extern "C" fn chg(params: api::hmac_finalize::Params) -> api::hmac_finali
 }
 
 #[no_mangle]
-unsafe extern "C" fn chr(_: api::is_hkdf_supported::Params) -> api::is_hkdf_supported::Results {
+unsafe extern "C" fn env_chr(_: api::is_hkdf_supported::Params) -> api::is_hkdf_supported::Results {
     api::is_hkdf_supported::Results { supported: 1 }
 }
 
 #[no_mangle]
-unsafe extern "C" fn che(params: api::hkdf_expand::Params) -> api::hkdf_expand::Results {
+unsafe extern "C" fn env_che(params: api::hkdf_expand::Params) -> api::hkdf_expand::Results {
     let api::hkdf_expand::Params { algorithm, prk, prk_len, info, info_len, okm, okm_len } = params;
     let prk = unsafe { std::slice::from_raw_parts(prk, prk_len) };
     let info = unsafe { std::slice::from_raw_parts(info, info_len) };