+++ crates/stub/src/debug.rs
@@ -15,7 +15,7 @@
 use wasefire_applet_api::debug as api;
 
 #[no_mangle]
-unsafe extern "C" fn dp(params: api::println::Params) {
+unsafe extern "C" fn env_dp(params: api::println::Params) {
     let api::println::Params { ptr, len } = params;
     let msg = unsafe { std::slice::from_raw_parts(ptr, len) };
     let msg = std::str::from_utf8(msg).unwrap();
@@ -23,7 +23,7 @@ unsafe extern "C" fn dp(params: api::println::Params) {
 }
 
 #[no_mangle]
-unsafe extern "C" fn de(params: api::exit::Params) {
+unsafe extern "C" fn env_de(params: api::exit::Params) {
     let api::exit::Params { code } = params;
     std::process::exit(code as i32)
 }