+++ crates/scheduler/CHANGELOG.md
@@ -81,4 +81,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 17 -->
+<!-- Increment to skip CHANGELOG.md test: 18 -->